+++ src/modules/newsFeed/NewsFeedTypes.ts
@@ -19,29 +19,12 @@ const NewsItemStatusLabel: Record<NewsItemStatus, string> = {
 export const getNewsItemStatusLabel = (status: NewsItemStatus): string =>
     NewsItemStatusLabel[status];
 
-// This order is the one defined in the backend
-enum TaskCategoryNumber {
-    train = 1,
-    aggregate = 2,
-    composite = 3,
-    test = 4,
+export enum NewsItemAssetKind {
+    computePlan = 'ASSET_COMPUTE_PLAN',
+    algo = 'ASSET_ALGO',
+    metric = 'ASSET_METRIC',
+    dataset = 'ASSET_DATA_MANAGER',
 }
-
-export const taskCategoryByCategoryNumber: Record<
-    TaskCategoryNumber,
-    TaskCategory
-> = {
-    [TaskCategoryNumber.train]: TaskCategory.train,
-    [TaskCategoryNumber.test]: TaskCategory.test,
-    [TaskCategoryNumber.composite]: TaskCategory.composite,
-    [TaskCategoryNumber.aggregate]: TaskCategory.aggregate,
-};
-
-export type NewsItemAssetKind =
-    | 'ASSET_COMPUTE_PLAN'
-    | 'ASSET_ALGO'
-    | 'ASSET_METRIC'
-    | 'ASSET_DATA_MANAGER';
 
 const NewsItemAssetLabel: Record<NewsItemAssetKind, string> = {
     ASSET_COMPUTE_PLAN: 'Compute plan',
@@ -60,7 +43,7 @@ export interface NewsItemType {
     status: NewsItemStatus;
     timestamp: string;
     detail: {
-        task_category: TaskCategoryNumber;
+        task_category: TaskCategory;
         first_failed_task_key?: string;
     };
 }