+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Newsfeed items link now redirecting to correct asset (#618)
+
 ## [0.26.0] - 2022-06-20
 
 ### Added